breakout files, apply style
