+++ vllm/model_executor/model_loader/reload/decorator.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+
+__all__ = ["supports_reloading"]
+
+
+def supports_reloading():
+    raise NotImplementedError()
