+++ vllm/model_executor/models/utils.py
@@ -22,7 +22,6 @@ from vllm.logger import init_logger
 from vllm.model_executor.layers.quantization.base_config import (
     QuantizationConfig,
 )
-from vllm.model_executor.model_loader.reload_utils import supports_reloading
 from vllm.model_executor.model_loader.weight_utils import default_weight_loader
 from vllm.model_executor.models.interfaces import supports_any_eagle
 from vllm.multimodal import NestedTensors
@@ -323,7 +322,7 @@ class AutoWeightsLoader:
                 )
                 raise ValueError(msg)
 
-    #@supports_reloading
+    # @supports_reloading
     def load_weights(
         self,
         weights: Iterable[tuple[str, torch.Tensor]],
